# chenoj-backend-microservice/chenoj-backend-question-service/src/main/java/com/chen/chenojbackendquestionservice/controller/inner/QuestionInnerController.java
package com.chen.chenojbackendquestionservice.controller.inner;

import com.chen.chenojbackendmodel.model.entity.Question;
import com.chen.chenojbackendmodel.model.entity.QuestionSubmit;
import com.chen.chenojbackendquestionservice.service.QuestionService;
import com.chen.chenojbackendquestionservice.service.QuestionSubmitService;
import com.chen.chenojbackendserviceclient.service.QuestionFeignClient;
import org.springframework.web.bind.annotation.*;

import javax.annotation.Resource;

/**
 * 该服务仅内部调用，不是给前端的
 */
@RestController
@RequestMapping("/inner")
public class QuestionInnerController implements QuestionFeignClient {

    @Resource
    private QuestionService questionService;

    @Resource
    private QuestionSubmitService questionSubmitService;


    @GetMapping("/get/id")
    @Override
    public Question getQuestionById(@RequestParam("questionId") long questionId) {
        return questionService.getById(questionId);
    }

    @GetMapping("/question_submit/get/id")
    @Override
    public QuestionSubmit getQuestionSubmitById(@RequestParam("questionId") long questionSubmitId) {
        return questionSubmitService.getById(questionSubmitId);
    }

    @PostMapping("/question_submit/update")
    @Override
    public boolean updateQuestionSubmitById(@RequestBody QuestionSubmit questionSubmit) {
        return questionSubmitService.updateById(questionSubmit);
    }
}