# chenoj-frontend/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "接口文档",
    "contact": {}
  },
  "host": "localhost:8121",
  "basePath": "/api",
...
  "paths": {
...
    "/api/question/get": {
      "get": {
        "tags": [
          "question-controller"
        ],
        "summary": "getQuestionById",
        "operationId": "getQuestionByIdUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "id",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«Question»",
              "$ref": "#/definitions/BaseResponse«Question»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«Question»",
              "$ref": "#/definitions/BaseResponse«Question»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
...
    "/api/question/question_submit/get/id": {
      "get": {
        "tags": [
          "question-controller"
        ],
        "summary": "getJudgeResult",
        "operationId": "getJudgeResultUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "id",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«QuestionSubmitVO»",
              "$ref": "#/definitions/BaseResponse«QuestionSubmitVO»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«QuestionSubmitVO»",
              "$ref": "#/definitions/BaseResponse«QuestionSubmitVO»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
...
  },
  "definitions": {
...
    "BaseResponse«QuestionSubmitVO»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "originalRef": "QuestionSubmitVO",
          "$ref": "#/definitions/QuestionSubmitVO"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "BaseResponse«QuestionSubmitVO»"
    },
...
    "BaseResponse«Question»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "originalRef": "Question",
          "$ref": "#/definitions/Question"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "BaseResponse«Question»"
    },
...
    "JudgeConfig": {
      "type": "object",
      "properties": {
        "memoryLimit": {
          "type": "integer",
          "format": "int64"
        },
        "stackLimit": {
          "type": "integer",
          "format": "int64"
        },
        "timeLimit": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "JudgeConfig"
    },
    "JudgeInfo": {
      "type": "object",
      "properties": {
        "memory": {
          "type": "integer",
          "format": "int64"
        },
        "message": {
          "type": "string"
        },
        "passNum": {
          "type": "integer",
          "format": "int32"
        },
        "time": {
          "type": "integer",
          "format": "int64"
        },
        "totalNum": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "JudgeInfo"
    },
...
    "Question": {
      "type": "object",
      "properties": {
        "acceptedNum": {
          "type": "integer",
          "format": "int32"
        },
        "answer": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "createTime": {
          "type": "string",
          "format": "date-time"
        },
        "favourNum": {
          "type": "integer",
          "format": "int32"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "isDelete": {
          "type": "integer",
          "format": "int32"
        },
        "judgeCase": {
          "type": "string"
        },
        "judgeConfig": {
          "type": "string"
        },
        "submitNum": {
          "type": "integer",
          "format": "int32"
        },
        "tags": {
          "type": "string"
        },
        "thumbNum": {
          "type": "integer",
          "format": "int32"
        },
        "title": {
          "type": "string"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time"
        },
        "userAccount": {
          "type": "string"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "Question"
    },
...
    "QuestionSubmitVO": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "createTime": {
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "judgeInfo": {
          "originalRef": "JudgeInfo",
          "$ref": "#/definitions/JudgeInfo"
        },
        "language": {
          "type": "string"
        },
        "questionId": {
          "type": "integer",
          "format": "int64"
        },
        "questionIdStr": {
          "type": "string"
        },
        "questionVO": {
          "originalRef": "QuestionVO",
          "$ref": "#/definitions/QuestionVO"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time"
        },
        "userAccount": {
          "type": "string"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        },
        "userVO": {
          "originalRef": "UserVO",
          "$ref": "#/definitions/UserVO"
        }
      },
      "title": "QuestionSubmitVO"
    },
...
    "QuestionVO": {
      "type": "object",
      "properties": {
        "acceptedNum": {
          "type": "integer",
          "format": "int32"
        },
        "answer": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "createTime": {
          "type": "string",
          "format": "date-time"
        },
        "favourNum": {
          "type": "integer",
          "format": "int32"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "judgeConfig": {
          "originalRef": "JudgeConfig",
          "$ref": "#/definitions/JudgeConfig"
        },
        "submitNum": {
          "type": "integer",
          "format": "int32"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "thumbNum": {
          "type": "integer",
          "format": "int32"
        },
        "title": {
          "type": "string"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        },
        "userVO": {
          "originalRef": "UserVO",
          "$ref": "#/definitions/UserVO"
        }
      },
      "title": "QuestionVO"
    },
...
    "UserVO": {
      "type": "object",
      "properties": {
        "createTime": {
          "type": "string",
          "format": "date-time"
        },
        "gender": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "userAvatar": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "userProfile": {
          "type": "string"
        },
        "userRole": {
          "type": "string"
        }
      },
      "title": "UserVO"
    }
  },
...
}